membagi file employee.java menjadi beberapa file
