+++ src/lib/spouse.java
@@ -1,0 +1,11 @@
+package lib;
+
+public class spouse extends Employee {
+    private String spouseName;
+	private String spouseIdNumber;
+
+    public void setSpouse(String spouseName, String spouseIdNumber) {
+		this.spouseName = spouseName;
+		this.spouseIdNumber = spouseIdNumber;
+	}
+}
